refactored express api creationg to app.ts
